cheaking the balance by the height diff left - right

+++ 14-binary_tree_balance.c
@@ -17,7 +17,7 @@ int binary_tree_balance(const binary_tree_t *tree)
 if (tree == NULL)
 return (0);
 
-return ((int)binary_trees_height(tree->left) - (int)binary_trees_height(tree->right));
+return (binary_trees_height(tree->left) - binary_trees_height(tree->right));
 }
 
 /**
@@ -32,9 +32,9 @@ return ((int)binary_trees_height(tree->left) - (int)binary_trees_height(tree->ri
  * Return:
  * The height of the tree as a size_t value.
  */
-size_t binary_trees_height(const binary_tree_t *tree)
+int binary_trees_height(const binary_tree_t *tree)
 {
-size_t l = 0, r = 0;
+int l = 0, r = 0;
 if (tree == NULL)
 return (-1);
