+++ binary_trees.h
@@ -39,7 +39,7 @@ size_t binary_tree_nodes(const binary_tree_t *tree);
 size_t binary_tree_siz(const binary_tree_t *tree);
 size_t binary_tree_leave(const binary_tree_t *tree);
 int binary_tree_balance(const binary_tree_t *tree);
-size_t binary_trees_height(const binary_tree_t *tree);
+int binary_trees_height(const binary_tree_t *tree);
 int binary_tree_is_full(const binary_tree_t *tree);
 int binary_tree_is_perfect(const binary_tree_t *tree);
 int height(const binary_tree_t *tree);
